refactor(MovieData): 패키지 변경
